requirements.txt
python-dotenv==1.1.0
uv==0.7.12
PyYAML==6.0.2
sentence-transformers==4.1.0
langchain-huggingface==0.3.1
langchain-community==0.3.24
faiss-gpu==1.7.2
faiss-cpu==1.11.0
google-search-results==2.4.2
accelerate==1.8.1
bitsandbytes==0.46.1
llama-cpp-python
protobuf==6.31.1
gradio
